DocIdSetBuilder code refactor
