+++ lucene/core/src/java/org/apache/lucene/util/DocIdSetBuilder.java
@@ -513,73 +513,6 @@ public final class DocIdSetBuilder {
   }
 
   /**
-   * Wrapper for partition-relative bitsets that offsets doc IDs back to absolute values when
-   * iterating.
-   */
-  private static class OffsetBitDocIdSet extends DocIdSet {
-    private final BitDocIdSet delegate;
-    private final int offset;
-
-    OffsetBitDocIdSet(FixedBitSet bitSet, long cost, int offset) {
-      this.delegate = new BitDocIdSet(bitSet, cost);
-      this.offset = offset;
-    }
-
-    @Override
-    public DocIdSetIterator iterator() {
-      DocIdSetIterator delegateIterator = delegate.iterator();
-      if (delegateIterator == null) {
-        return null;
-      }
-      return new OffsetDocIdSetIterator(delegateIterator, offset);
-    }
-
-    @Override
-    public long ramBytesUsed() {
-      return delegate.ramBytesUsed();
-    }
-  }
-
-  /**
-   * Iterator that adds an offset to all doc IDs from the underlying iterator, converting
-   * partition-relative indices back to absolute doc IDs.
-   */
-  private static class OffsetDocIdSetIterator extends DocIdSetIterator {
-    private final DocIdSetIterator delegate;
-    private final int offset;
-
-    OffsetDocIdSetIterator(DocIdSetIterator delegate, int offset) {
-      this.delegate = delegate;
-      this.offset = offset;
-    }
-
-    @Override
-    public int docID() {
-      int doc = delegate.docID();
-      return doc == NO_MORE_DOCS ? NO_MORE_DOCS : doc + offset;
-    }
-
-    @Override
-    public int nextDoc() throws IOException {
-      int doc = delegate.nextDoc();
-      return doc == NO_MORE_DOCS ? NO_MORE_DOCS : doc + offset;
-    }
-
-    @Override
-    public int advance(int target) throws IOException {
-      // Convert target from absolute to partition-relative, advance, then convert back
-      int relativeTarget = target - offset;
-      int doc = delegate.advance(Math.max(0, relativeTarget));
-      return doc == NO_MORE_DOCS ? NO_MORE_DOCS : doc + offset;
-    }
-
-    @Override
-    public long cost() {
-      return delegate.cost();
-    }
-  }
-
-  /**
    * Concatenate the buffers in any order, leaving at least one empty slot in the end NOTE: this
    * method might reuse one of the arrays
    */
